+++ .prettierignore
@@ -1,0 +1,6 @@
+node_modules/
+dist/
+
+LICENSE
+package-lock.json
+tsconfig.tsbuildinfo
