+++ coyote_html/package.json
@@ -5,7 +5,8 @@
 	"type": "module",
 	"main": "dist/mod.js",
 	"scripts": {
-		"build": "tsc --build"
+		"build": "npx tsc --build",
+		"test": ""
 	},
 	"author": "wolfpup",
 	"license": "BSD-3-Clause"
