+++ html/src/compose.test.ts
@@ -1,0 +1,38 @@
+import { compose } from "./mod.js";
+import { Sieve, ClientSieve } from "../../sieve/dist/mod.js";
+
+function testCompose() {
+	const expected = "";
+	const templateStr = `
+		<form>
+			<input>
+			<div>
+			</div>
+			<p>you're a boy kisser, aren't you</p>
+			<input type=checkbox>
+		</form>
+	`;
+	const sieve = new Sieve();
+	// const sieve = new ClientSieve();
+	const coyoteTemplate = compose(sieve, templateStr);
+
+	if (expected !== coyoteTemplate) {
+		return `
+expected:
+${expected}
+		
+found:
+${coyoteTemplate}`
+	}
+
+	return;
+}
+
+export const tests = [
+	testCompose,
+];
+
+export const options = {
+	title: import.meta.url,
+	runAsynchronously: false,
+}
