+++ html/src/mod.test.ts
@@ -1,0 +1,5 @@
+import * as composeTests from "./compose.test.js";
+
+export const testModules = [
+	composeTests,
+];
