+++ package.json
@@ -6,11 +6,14 @@
         "coyote_html"
     ],
     "scripts": {
-        "build": "npm run --workspaces build"
+        "build": "npm run --workspaces build",
+        "test": "npm run --workspaces test",
+        "fmt": "npx prettier --write ./"
     },
     "devDependencies": {
         "@types/node": "^22.5.5",
         "jackrabbit": "github:wolfpup-software/jackrabbit-js",
+        "prettier": "^3.3.3",
         "typescript": "^5.6.3"
     }
 }
