+++ parse/src/mod.test.ts
@@ -1,0 +1,5 @@
+import * as parseStrTests from "./parse_str.test.js";
+
+export const testModules = [
+	parseStrTests,
+];
